centralize mixtral api over replicate
