+++ src/commands/ai/replicate.ts
@@ -4,3 +4,7 @@ import Replicate from 'replicate';
 export const replicate = new Replicate({
     auth: ENV.REPLICATE_TOKEN,
 });
+
+export function check_authentication(): boolean {
+    return !!replicate.auth;
+}
